Corrige deserialização JSON de entidades relacionadas

+++ app/src/main/java/br/com/fiap/easycoachapp/domain/entities/SessionEntity.kt
@@ -6,19 +6,20 @@ import java.util.Date
 class SessionEntity (
     val uid: String,
     val scheduledDateTime: Date,
-    val sessionNumber: Int,
+    val sessionNumber: Long,
     val title: String,
     val description: String?,
     val inviteUrl: String?,
     val hasCancellationFee: Boolean,
     val specialtyUid: String
-) {
+)
+{
     companion object {
         fun fromJson(json: MutableMap<String, Any>) : SessionEntity {
             return SessionEntity(
                 uid = json["uid"].toString(),
                 scheduledDateTime = (json["scheduledDateTime"] as Timestamp).toDate(),
-                sessionNumber = json["sessionNumber"] as Int,
+                sessionNumber = json["sessionNumber"] as Long,
                 title = json["title"].toString(),
                 description = json["description"]?.toString(),
                 inviteUrl = json["inviteUrl"].toString(),
